requirements.txt
albumentations==1.3.1
imageio==2.31.5
matplotlib==3.7.3
MedPy==0.4.0
monai==1.3.0
nibabel==5.1.0
numpy==1.23.1
opencv-python==4.8.1.78
pandas==2.0.3
Pillow==10.1.0
PyYAML==6.0.1
pyzmq==25.1.1
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
SimpleITK==2.3.0
tensorboardX==2.6.2.2
timm==0.9.7
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
